rename: 메인페이지 컴포넌트 폴더구조 변경
