+++ src/pages/Main/components/SearchBar/style.scss
