+++ src/pages/Main/components/SpaceCard/style.scss
