+++ src/pages/Main/components/SpaceCardList/style.scss
